+++ copl/com.elster.jupiter.hsm/src/test/java/com/elster/jupiter/hsm/model/request/ImportKeyRequestIT.java
@@ -1,19 +1,17 @@
 package com.elster.jupiter.hsm.model.request;
 
-import com.elster.jupiter.hsm.impl.config.HsmConfigurationPropFileImpl;
-import com.elster.jupiter.hsm.model.HsmBaseException;
-import com.elster.jupiter.hsm.impl.config.HsmConfiguration;
-import com.elster.jupiter.hsm.model.keys.HsmKeyType;
-import com.elster.jupiter.hsm.model.keys.SessionKeyCapability;
-import com.elster.jupiter.hsm.model.krypto.AsymmetricAlgorithm;
-import com.elster.jupiter.hsm.model.krypto.SymmetricAlgorithm;
-
 import com.atos.worldline.jss.api.custom.energy.AESDeviceKey;
 import com.atos.worldline.jss.api.custom.energy.DeviceKey;
 import com.atos.worldline.jss.api.custom.energy.TransportKey;
 import com.atos.worldline.jss.api.key.KeyLabel;
 import com.atos.worldline.jss.api.key.UnsupportedKEKEncryptionMethodException;
-
+import com.elster.jupiter.hsm.impl.config.HsmConfiguration;
+import com.elster.jupiter.hsm.impl.config.HsmConfigurationPropFileImpl;
+import com.elster.jupiter.hsm.model.HsmBaseException;
+import com.elster.jupiter.hsm.model.keys.HsmKeyType;
+import com.elster.jupiter.hsm.model.keys.SessionKeyCapability;
+import com.elster.jupiter.hsm.model.krypto.AsymmetricAlgorithm;
+import com.elster.jupiter.hsm.model.krypto.SymmetricAlgorithm;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -30,7 +28,7 @@ public class ImportKeyRequestIT {
     private static final byte[] DEVICE_KEY = new byte[]{9,8,7,6,5};
     private static final byte[] DEVICE_KEY_INIT_VECTOR = new byte[]{5,4,3,2,1};;
     public static final int KEY_SIZE = 32;
-    private static final HsmKeyType HSM_KEY_TYPE = new HsmKeyType(LABEL, SessionKeyCapability.SM_KEK_AGREEMENT, SessionKeyCapability.SM_KEK_RENEWAL, (short) KEY_SIZE);
+    private static final HsmKeyType HSM_KEY_TYPE = new HsmKeyType(LABEL, SessionKeyCapability.SM_KEK_AGREEMENT, SessionKeyCapability.SM_KEK_RENEWAL, KEY_SIZE);
 
     private static HsmConfiguration HSM_CONFIG;
     private static ImportKeyRequest IKR;
